pyproject.toml
[tool.poetry]
name = "hakaton-server"
version = "0.1.0"
description = ""
authors = ["hakaton"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.114.0"}
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
